Add files via upload
Pokémon webapp is a simple React-based project that fetches data from the PokeAPI and displays a list of Pokémon with their images. On clicking a Pokémon, detailed information like height, weight, and an official artwork image is shown. A "Back" button allows users to return to the full list.